% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbin1.r
\name{rcbin1}
\alias{rcbin1}
\title{Generates correlated binary cluster data}
\usage{
rcbin1(prop = 0.5, prvar = 0, noc, csize, csvar = 0, mincsize = 2,
  rho)
}
\arguments{
\item{prop}{A numeric value between 0 and 1 denoting assumed proportion of event in interest, default value is 0.5. See Detail}

\item{prvar}{A numeric value between 0 and 1 denoting varince in assumed proportion of event (\code{prvar}), default value is 0. See Detail}

\item{noc}{A positive numeric value telling the number of clusters to be generated}

\item{csize}{A numeric value (\eqn{\ge 2}) denoting cluster size desired}

\item{csvar}{A positive numeric value denoting Variance of cluster size, default value is 0, see Detail}

\item{mincsize}{A numeric value (\eqn{\ge 2}) denoting the minimum cluster size desired, default value is 2, see Detail}

\item{rho}{A numeric value between 0 and 1 denoting desired level of Intracluster Correlation}
}
\value{
A dataframe with two columns presenting cluster id (\code{cid}) and a binary response (\code{y}) variables
}
\description{
Generates correrlated binary cluster data given value of Intracluster Correlation, proportion of event and it's variance, number of clusters, cluster size and it's variance, and minimum cluster size
}
\details{
If supplied value of \code{prvar} is 0, the event proportion for all clusters is considered constant as supplied by \code{prop}.
If supplied \code{prvar} is > 0, cluster specific event proportions are generated from Beta distribution with
\code{shape1} and \code{shape2} parameters \eqn{a} and \eqn{b} respectively, see \code{\link{rbeta}}
The shape parameters are obtained using supplied values of \code{prop} and \code{prvar} by solving the equations
 \code{prop} \eqn{= a/(a + b)} and \code{prvar} \eqn{= ab/[(a + b)^2(1 + a + b)]}

If supplied value of \code{csvar} is 0, cluster of equal size (\code{csize}) will be generated. For \code{csvar} > 0, will be generated from
Normal or Negative Binomial dsitributions depending on relationship between \code{csize} and \code{csvar}.
If \code{csvar} < \code{csize}, the varying cluster sizes will be generated
from a Normal distribution with mean = \code{csize} and variacne = \code{csvar} (see \code{\link{rnorm}}).
If \code{csvar} \eqn{\ge} \code{csize} i.e. in the case of overdispersion,
cluster sizes will be generated from Negative Bionomial distribution using \code{mu} = \code{csize} and
\ifelse{latex}{\eqn{\code{size} = \code{csize}/[\code{csize}(\code{cscv}^2 - 1)]}}{\code{size} = \code{csize}/(\code{csize}*(\code{cscv}^2 - 1))}
(see \code{\link{rnbinom}}), where \code{cscv} is the coefficient of variation of cluster sizes defined as
\ifelse{latex}{\eqn{\sqrt{\code{csvar}}/\code{csize}}}{\code{sqrt(csvar)}/\code{csize}}. If the size of any cluster
is generated as less than 2, it will be replaced by the supplied value of minimum cluster size (\code{mincsize}) which has a default value
of 2
}
\examples{
rcbin1(prop = .6, prvar = .1, noc = 100, csize = 10, csvar = 12, rho = 0.2, mincsize = 2)

}
\references{
Lunn, A.D. and Davies, S.J., 1998. A note on generating correlated binary variables. Biometrika, 85(2), pp.487-490.
}
\seealso{
\code{\link{rcbin}} \code{\link{iccbin}}
}
\author{
Akhtar Hossain \email{mhossain@email.sc.edu}
}
